12. Numbers have life; they're not just symbols on paper

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -6,9 +6,9 @@
  * Return: nothing.
  */
 
-void print_number(int n)
+void print_number(long int n)
 {
-	int divisor, digit;
+	long int divisor, digit;
 
 	divisor = 1;
